+++ dtoolcore/storagebroker.py
@@ -116,6 +116,16 @@ class DiskStorageBroker(object):
         with open(self._manifest_abspath) as fh:
             return json.load(fh)
 
+
+    def list_overlay_names(self):
+        """Return list of overlay names."""
+        overlay_names = []
+        for fname in os.listdir(self._overlays_abspath):
+            name, ext = os.path.splitext(fname)
+            overlay_names.append(name)
+        return overlay_names
+
+
     def get_overlay(self, overlay_name):
         """Return overlay as a dictionary.